Slight changes to Solo output to perfectly match CR2 and very close match to CR3.1.0

+++ source/ParametersSolo.h
@@ -56,7 +56,7 @@ public:
     
     //filtering
     char QSbase,QSmax;//quality score base and cutoff
-    float cbMinP;//for CBs with non-exact matching to WL, min posterior probability
+    double cbMinP;//for CBs with non-exact matching to WL, min posterior probability
     
     //cell filtering
     struct {
